venture_ai_core/requirements.txt
langchain
langgraph
langchain_google_genai
python-dotenv
requests
openai